+++ R/read.R
@@ -336,9 +336,25 @@ get_data_units = function(data) {
 	}
 }
 
+#' Read data using GDAL's multidimensional array API (experimental)
+#' 
+#' Read data using GDAL's multidimensional array API (experimental)
+#' @param x data source name
+#' @param variable name of the array to be read
+#' @param options array opening options
+#' @param raster names of the raster variables (default: first two)
+#' @param offset integer; offset for each dimension (pixels) of sub-array to read (default: 0,0,0,...) (requires sf >= 1.0-9)
+#' @param count integer; size for each dimension (pixels) of sub-array to read (default: read all) (requires sf >= 1.0-9)
+#' @param step integer; step size for each dimension (pixels) of sub-aray to read (requires sf >= 1.0-9)
+#' @details it is assumed that the first two dimensions are easting / northing
+#' @param ... ignored
+#' @export
 read_mdim = function(x, variable = character(0), ..., options = character(0), raster = NULL,
 					 offset = integer(0), count = integer(0), step = integer(0)) {
-	ret = gdal_read_mdim(x, variable, options, offset, count, step)
+	ret = if (packageVersion("sf") >= "1.0-9")
+			gdal_read_mdim(x, variable, options, rev(offset), rev(count), rev(step))
+		else
+			gdal_read_mdim(x, variable, options)
 	create_units = function(x) {
 		u <- attr(x, "units")
 		if (is.null(u) || u == "")
@@ -357,10 +373,30 @@ read_mdim = function(x, variable = character(0), ..., options = character(0), ra
 			}
 		}
 	}
-	l = lapply(ret$dimensions, function(x) create_units(x$values[[1]]))
-	d = rev(lapply(l, function(x) create_dimension(values = x)))
+	l = rev(lapply(ret$dimensions, function(x) create_units(x$values[[1]])))
+	if (length(offset) != 0 || length(step) != 0 || length(count) != 0) {
+		if (length(offset) == 0)
+			offset = rep(0, length(l))
+		if (length(step) == 0)
+			step = rep(1, length(l))
+		if (length(count) == 0)
+			count = floor((lengths(l) - offset)/step)
+		for (i in seq_along(l)) {
+			l[[i]] = l[[i]][seq(from = offset[i]+1, length.out = count[i], by = step[i])]
+		}
+	}
+	d = mapply(function(x, i) create_dimension(values = x, is_raster = i %in% 1:2), l, seq_along(l),
+			SIMPLIFY = FALSE)
 	if (is.null(raster))
 		raster = get_raster(dimensions = names(d)[1:2])
+	else
+		raster = get_raster(dimensions = raster)
+
+	# handle array units:
+	for (i in seq_along(ret$array_list))
+		if (nchar(u <- attr(ret$array_list[[i]], "units")))
+			ret$array_list[[i]] = units::set_units(ret$array_list[[i]], u, mode = "standard")
+
 	lst = lapply(ret$array_list, function(x) structure(x, dim = rev(dim(x))))
 	st_set_crs(st_stars(lst, dimensions = structure(d, raster = raster, class = "dimensions")),
 		ret$srs)